Make crate useable as a library
